+++ services/web/server/tests/unit/with_dbs/04/studies_dispatcher/test_studies_dispatcher_rest_redirects.py
@@ -11,7 +11,6 @@ from typing import Any
 from unittest import mock
 
 import pytest
-import simcore_service_webserver.studies_dispatcher
 import sqlalchemy as sa
 from aiohttp import ClientResponse, ClientSession
 from aiohttp.test_utils import TestClient, TestServer
@@ -20,23 +19,16 @@ from common_library.json_serialization import json_dumps
 from common_library.serialization import model_dump_with_secrets
 from common_library.users_enums import UserRole
 from models_library.projects_state import ProjectShareState, ProjectStatus
-from pydantic import BaseModel, ByteSize, TypeAdapter
+from pydantic import ByteSize, TypeAdapter
 from pytest_mock import MockerFixture
 from pytest_simcore.helpers.assert_checks import assert_status
 from pytest_simcore.helpers.monkeypatch_envs import setenvs_from_dict
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from pytest_simcore.helpers.webserver_users import UserInfoDict
-from pytest_simcore.pydantic_models import (
-    assert_validation_model,
-    walk_model_examples_in_package,
-)
 from servicelib.aiohttp import status
 from settings_library.rabbit import RabbitSettings
 from settings_library.redis import RedisSettings
 from settings_library.utils_session import DEFAULT_SESSION_COOKIE_NAME
-from simcore_service_webserver.studies_dispatcher._controller.rest.nih_schemas import (
-    ServiceGet,
-)
 from simcore_service_webserver.studies_dispatcher._models import ViewerInfo
 from sqlalchemy.sql import text
 from yarl import URL
@@ -181,123 +173,6 @@ FAKE_VIEWS_LIST = [
         input_port_key="input_1",
     ),
 ]
-
-
-# REST-API
-#  Samples taken from trials on http://127.0.0.1:9081/dev/doc#/viewer/get_viewer_for_file
-#
-
-
-def _get_base_url(client: TestClient) -> str:
-    s = client.server
-    assert isinstance(s.scheme, str)
-    url = URL.build(scheme=s.scheme, host=s.host, port=s.port)
-    return f"{url}"
-
-
-async def test_api_get_viewer_for_file(client: TestClient):
-    resp = await client.get("/v0/viewers/default?file_type=JPEG")
-    data, _ = await assert_status(resp, status.HTTP_200_OK)
-
-    base_url = _get_base_url(client)
-    assert data == [
-        {
-            "file_type": "JPEG",
-            "title": "Bio-formats v1.0.1",
-            "view_url": f"{base_url}/view?file_type=JPEG&viewer_key=simcore/services/dynamic/bio-formats-web&viewer_version=1.0.1",
-        },
-    ]
-
-
-async def test_api_get_viewer_for_unsupported_type(client: TestClient):
-    resp = await client.get("/v0/viewers/default?file_type=UNSUPPORTED_TYPE")
-    data, error = await assert_status(resp, status.HTTP_200_OK)
-    assert data == []
-    assert error is None
-
-
-async def test_api_list_supported_filetypes(client: TestClient):
-    resp = await client.get("/v0/viewers/default")
-    data, _ = await assert_status(resp, status.HTTP_200_OK)
-
-    base_url = _get_base_url(client)
-    assert data == [
-        {
-            "title": "Rawgraphs v2.11.1",
-            "file_type": "CSV",
-            "view_url": f"{base_url}/view?file_type=CSV&viewer_key=simcore/services/dynamic/raw-graphs&viewer_version=2.11.1",
-        },
-        {
-            "title": "Jupyterlab math v1.6.9",
-            "file_type": "IPYNB",
-            "view_url": f"{base_url}/view?file_type=IPYNB&viewer_key=simcore/services/dynamic/jupyter-octave-python-math&viewer_version=1.6.9",
-        },
-        {
-            "title": "Bio-formats v1.0.1",
-            "file_type": "JPEG",
-            "view_url": f"{base_url}/view?file_type=JPEG&viewer_key=simcore/services/dynamic/bio-formats-web&viewer_version=1.0.1",
-        },
-        {
-            "title": "Rawgraphs v2.11.1",
-            "file_type": "JSON",
-            "view_url": f"{base_url}/view?file_type=JSON&viewer_key=simcore/services/dynamic/raw-graphs&viewer_version=2.11.1",
-        },
-        {
-            "title": "Bio-formats v1.0.1",
-            "file_type": "PNG",
-            "view_url": f"{base_url}/view?file_type=PNG&viewer_key=simcore/services/dynamic/bio-formats-web&viewer_version=1.0.1",
-        },
-        {
-            "title": "Jupyterlab math v1.6.9",
-            "file_type": "PY",
-            "view_url": f"{base_url}/view?file_type=PY&viewer_key=simcore/services/dynamic/jupyter-octave-python-math&viewer_version=1.6.9",
-        },
-        {
-            "title": "Rawgraphs v2.11.1",
-            "file_type": "TSV",
-            "view_url": f"{base_url}/view?file_type=TSV&viewer_key=simcore/services/dynamic/raw-graphs&viewer_version=2.11.1",
-        },
-        {
-            "title": "Rawgraphs v2.11.1",
-            "file_type": "XLSX",
-            "view_url": f"{base_url}/view?file_type=XLSX&viewer_key=simcore/services/dynamic/raw-graphs&viewer_version=2.11.1",
-        },
-    ]
-
-
-@pytest.mark.parametrize(
-    "model_cls, example_name, example_data",
-    walk_model_examples_in_package(simcore_service_webserver.studies_dispatcher),
-)
-def test_model_examples(
-    model_cls: type[BaseModel], example_name: str, example_data: Any
-):
-    assert_validation_model(
-        model_cls, example_name=example_name, example_data=example_data
-    )
-
-
-async def test_api_list_services(client: TestClient):
-    assert client.app
-
-    url = client.app.router["list_latest_services"].url_for()
-    response = await client.get(f"{url}")
-
-    data, error = await assert_status(response, status.HTTP_200_OK)
-
-    services = TypeAdapter(list[ServiceGet]).validate_python(data)
-    assert services
-
-    # latest versions of services with everyone + ospar-product (see stmt_create_services_access_rights)
-    assert services[0].key == "simcore/services/dynamic/raw-graphs"
-    assert services[0].file_extensions == ["CSV", "JSON", "TSV", "XLSX"]
-    assert "2.11.1" in services[0].view_url.query
-
-    assert services[1].key == "simcore/services/dynamic/jupyter-octave-python-math"
-    assert services[1].file_extensions == ["IPYNB", "PY"]
-    assert "1.6.9" in services[1].view_url.query
-
-    assert error is None
 
 
 # REDIRECT ROUTES --------------------------------------------------------------------------------
